requirements.txt
tensorflow==2.15.1
numpy==1.26.4
opencv-python-headless==4.8.1.78
streamlit==1.53.0
streamlit-drawable-canvas==0.9.3
pillow
